Fix sfn (#107)
* lte: Provide correct Fr/Sf number to the methods computing the absolute Fr and Sf number

All the methods computing the absolute subframe number, specifically, the one in LteSlPool class, assume that the frame and subframe number range from 0 to 1023. However, in the simulator they range from 1 to 1024. Therefore, it is very important to subtract 1 frame and subframe number before calling these methods, and then add 1 in the resultant frame and subframe number.

* lte: Change m_ChangeOfSyncRefTrace to m_changeOfSyncRefTrace

Follow ns-3 variable name style guideline

* lte: Place logging statement after computing the offset for SLSS TX

* lte: Update the documentation of already documented examples

* lte: Add new SL synchronization test and documentation

* lte: Update sidelink doc pdf